+++ demo/integration-tests/webapp-vaadin-it/src/test/java/software/xdev/sse/demo/vaadin/cases/LoginOIDCTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.rnorth.ducttape.unreliables.Unreliables;
 
 import software.xdev.sse.demo.entities.UserDetail;
 import software.xdev.sse.demo.persistence.jpa.dao.UserDetailDAO;
@@ -60,7 +61,7 @@ class LoginOIDCTest extends InfraPerCaseTest
 		this.navigateTo("another");
 		
 		// Delete all cookies of the CURRENT domain
-		this.getWebDriver().manage().deleteAllCookies();
+		this.ensureAllCookiesDeleted();
 		this.getWebDriver().navigate().refresh();
 		
 		// Should be restored to the same path/view
@@ -79,12 +80,22 @@ class LoginOIDCTest extends InfraPerCaseTest
 		this.navigateTo("anotherThatDoesNotExist");
 		
 		// Delete all cookies of the CURRENT domain
-		this.getWebDriver().manage().deleteAllCookies();
+		this.ensureAllCookiesDeleted();
 		this.getWebDriver().navigate().refresh();
 		
 		// Should be restored to the same path/view
 		Assertions.assertDoesNotThrow(() ->
 			this.waitUntil(ExpectedConditions.urlToBe(this.getWebAppBaseUrl() + "/main")));
+	}
+	
+	private void ensureAllCookiesDeleted()
+	{
+		// Retry because GHA machine is sometimes failing here
+		Unreliables.retryUntilSuccess(
+			3, () -> {
+				this.getWebDriver().manage().deleteAllCookies();
+				return this.getWebDriver().manage().getCookies().isEmpty();
+			});
 	}
 	
 	@DisplayName("Check Login works when OIDC offline")