+++ web-sidecar-actuator/src/main/java/software/xdev/sse/web/sidecar/actuator/auto/ActuatorWebSecurityAutoConfig.java
@@ -15,6 +15,7 @@
  */
 package software.xdev.sse.web.sidecar.actuator.auto;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.actuate.autoconfigure.endpoint.web.WebEndpointProperties;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
@@ -28,11 +29,14 @@ import io.micrometer.core.instrument.MeterRegistry;
 import software.xdev.sse.web.sidecar.actuator.ActuatorBlackHolingPathsProvider;
 import software.xdev.sse.web.sidecar.actuator.ActuatorWebSecurity;
 import software.xdev.sse.web.sidecar.actuator.config.ActuatorSecurityConfig;
+import software.xdev.sse.web.sidecar.actuator.httpsecurity.ActuatorHttpSecMCustomizerContainer;
 import software.xdev.sse.web.sidecar.actuator.metrics.ActuatorSecurityMetricsHandler;
 import software.xdev.sse.web.sidecar.actuator.metrics.DefaultActuatorSecurityMetricsHandler;
 import software.xdev.sse.web.sidecar.actuator.passwordhash.cache.PasswordHashCache;
 import software.xdev.sse.web.sidecar.actuator.passwordhash.cache.UnchachedPasswordHashCache;
 import software.xdev.sse.web.sidecar.actuator.passwordhash.hasher.sha256.DefaultSHA256PasswordHasher;
+import software.xdev.sse.web.sidecar.httpsecurity.HttpSecurityMatcherPatternApplier;
+import software.xdev.sse.web.sidecar.httpsecurity.HttpSecurityMatcherPatternCreator;
 
 
 @ConditionalOnProperty(value = "sse.sidecar.actuator.enabled", matchIfMissing = true)
@@ -81,5 +85,17 @@ public class ActuatorWebSecurityAutoConfig
 	public PasswordHashCache passwordHashCache()
 	{
 		return new UnchachedPasswordHashCache();
+	}
+	
+	@ConditionalOnProperty(
+		value = "sse.sidecar.actuator.http-security-matcher.default.enabled",
+		matchIfMissing = true)
+	@ConditionalOnMissingBean
+	@Bean
+	public ActuatorHttpSecMCustomizerContainer actuatorHttpSecMCustomizerContainer(
+		final HttpSecurityMatcherPatternApplier applier,
+		@Autowired(required = false) final HttpSecurityMatcherPatternCreator creator)
+	{
+		return new ActuatorHttpSecMCustomizerContainer(applier, creator);
 	}
 }