+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 1.5.0
+* Vaadin
+  * Made the way `HttpSecurity#securityMatcher` is applied in Sidecars customizable #221
+    * By default `PathPatternRequestMatcher` is now used instead of `MvcRequestMatcher` or `AntPathRequestMatcher`
+    * This fixes problems where the app/servlet utilizes an existing `urlMapping` that could result in incorrect path interception.
+    * Further details can be found in `software.xdev.sse.web.sidecar.httpsecurity`
+  * Fix incorrect lookup of `LoginUrlStore` in `TotalVaadinFlowSecurityConfigurer#configureLoginViewFromLoginUrlStore`
+  * `TotalVaadinFlowSecurityConfigurer` make it possible to enable/disable certain configurations
+* Web
+  * Add customizer for [HSTS](https://en.wikipedia.org/wiki/HTTP_Strict_Transport_Security)
+    * also used by sidecars (e.g. Actuator)
+    * HSTS is disabled by default if SSL/HTTPS is also disabled
+    * Further details can be found in `software.xdev.sse.web.hsts`
+* Updated dependencies
+
 # 1.4.0
 * Vaadin
   * Deprecated `TotalVaadinFlowWebSecurity` because `VaadinWebSecurity` is deprecated